Create README.md
# EDA-For-Zoom-Stock-in-Python
This repository is an exploratory data analysis (EDA) notebook on the stock, Zoom. I wanted to use Zoom since it had a meteoric rise in price during the COVID pandemic. 

In this EDA, will share some tips to help explore any stock with simple libraries from python such as datetime search criterias with resample and the following funciton to explore the rolling mean of some of the zoom stock characteristics.

In this EDA, I use the following libraries (modules used in parenthesis):

  1. datetime (datetime (not a mistake! The module's name is the same as the library name.)
  2. pandas
  3. pandas_datareader
  4. matplotlib.pyplot ("inline" to help with subplots)

Feel feel to reach out with any questions!